+++ src/base/mem.c
@@ -67,16 +67,22 @@ static void* tk_alloc_impl(uint32_t size) {
 static void* tk_realloc_impl(void* ptr, uint32_t size) {
   if (ptr != NULL) {
     mem_block_t* head = (mem_block_t*)((char*)ptr - sizeof(mem_block_t));
-    void* newptr = tk_alloc_impl(size);
+    uint32_t old_size = head->size - sizeof(mem_block_t);
 
-    if (newptr) {
-      memcpy(newptr, ptr, tk_min(size, head->size - sizeof(mem_block_t)));
-      tk_free_impl(ptr);
-
-      return newptr;
-    } else {
+    if (size <= old_size) {
       return ptr;
+    } else {
+      void* newptr = tk_alloc_impl(size);
+      if (newptr) {
+        memcpy(newptr, ptr, tk_min(size, old_size));
+        tk_free_impl(ptr);
+
+        return newptr;
+      } else {
+        return ptr;
+      }
     }
+
   } else {
     return tk_alloc_impl(size);
   }