+++ src/base/style_factory.h
@@ -26,7 +26,46 @@
 
 BEGIN_C_DECLS
 
-style_t* style_factory_create_style(widget_t* widget);
+struct _style_factory_t;
+typedef struct _style_factory_t style_factory_t;
+
+typedef style_t* (*style_factory_create_style_t)(style_factory_t* factory, widget_t* widget);
+
+/**
+ * @class style_factory_t
+ * style_factory接口。
+ */
+struct _style_factory_t {
+  style_factory_create_style_t create_style;
+};
+
+/**
+ * @method style_factory_create_style
+ * 创建style。
+ * @param {style_factory_t*} factory factory对象。
+ * @param {widget_t*} widget 控件对象。
+ *
+ * @return {style_t*} 返回style对象。
+ */
+style_t* style_factory_create_style(style_factory_t* factory, widget_t* widget);
+
+/**
+ * @method style_factory
+ * 获取缺省的style factory。
+ * @alias style_factory_instance
+ * @annotation ["constructor"]
+ * @return {style_factory_t*} 返回style factory对象。
+ */
+style_factory_t* style_factory(void);
+
+/**
+ * @method style_factory_set
+ * 设置缺省的style factory(需要自定义的style factory才设置)。
+ * @param {style_factory_t*}  factory factory对象。
+ *
+ * @return {ret_t} 返回RET_OK表示成功，否则表示失败。
+ */
+ret_t style_factory_set(style_factory_t* factory);
 
 END_C_DECLS